Split serial_msp.c into serial and MSP parts (#571)
* Split serial_msp.c into serial and MSP parts
* Fixup after rebase
